view friends, user count and user info when clicked
