+++ client/src/components/Loader.js
@@ -1,0 +1,10 @@
+import React from "react";
+import "./Loader.css";
+
+export default function Loader() {
+  return (
+    <div className="spinner-container">
+      <div className="loading-spinner"></div>
+    </div>
+  );
+}
