+++ client/src/router/router.js
@@ -2,6 +2,9 @@ import React, { lazy, Suspense } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { user } from "../useGun";
 
+// spinner
+import Loader from "../components/Loader";
+
 // route elements (pages)
 const LandingPage = lazy(() => import("../pages/LandingPage"));
 const SignIn = lazy(() => import("../pages/SignIn"));
@@ -11,7 +14,7 @@ const Homepage = lazy(() => import("../pages/Homepage"));
 export default function Router() {
   return (
     <BrowserRouter>
-      <Suspense fallback={<>Loading... </>}>
+      <Suspense fallback={<Loader />}>
         <Routes>
           <Route path="/" element={user.is ? <Homepage /> : <LandingPage />} />
           <Route path="/login" element={<SignIn />} />
